requirements.txt
# Core Python packages
torch>=2.0.0
transformers>=4.52.0
pillow>=9.0.0
markdown-table>=0.0.8
pydantic>=2.0.0
tqdm>=4.65.0

# Optional: for Azure/OpenAI LLM integration
openai>=1.99.9,<2.0.0
langchain-openai>=0.3.32
langchain-core>=0.3.74,<1.0.0

# Optional: for PDF processing / OCR
PyMuPDF>=1.22.0
pdfplumber>=0.9.0
pytesseract>=0.3.10

# Optional: for image and table processing
numpy>=1.24.0
pandas>=2.0.0

# Optional: for Hugging Face model downloads
huggingface-hub>=0.17.0

# Any other utilities
requests>=2.32.0
